fix exception
